refactor : Move ClusterAccess and ClusterConfiguration to `jkube-kit/common`
+ Move ClusterAccess and ClusterConfiguration classes from
  `jkube-kit/config/resource` -> `jkube-kit/common` so that they're
  accessible to JKubeConfiguration
+ Add new field `clusterConfiguration` in JKubeConfiguration for passing
  in Cluster configuration
+ Add new field `currentContext` to ClusterConfiguration in order to
  pass current context
