+++ jkube-kit/common/src/main/java/org/eclipse/jkube/kit/common/access/ClusterAccess.java
@@ -11,7 +11,7 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-package org.eclipse.jkube.kit.config.access;
+package org.eclipse.jkube.kit.common.access;
 
 
 import io.fabric8.kubernetes.client.Config;
